significant refactor
