+++ src/lib/ezbot.spec.ts
@@ -5,19 +5,17 @@
 /* eslint-disable functional/no-return-void */
 import { trackPageView } from '@snowplow/browser-tracker';
 import { BrowserTracker } from '@snowplow/browser-tracker-core';
-import Ajv from "ajv";
+import Ajv from 'ajv';
 
-import {
-  initEzbot,
-  PredictionsContext,
-  startActivityTracking,
-  trackRewardEvent
-} from './ezbot';
-import * as predictions_schema from './schemas/com.ezbot/predictions_context/jsonschema/1-0-1.json';
-
+import { initEzbot } from './ezbot';
+import * as predictionsContextSchema from './schemas/com.ezbot/predictions_context/jsonschema/1-0-1.json';
+import { startActivityTracking, trackRewardEvent } from './tracking';
+import { EzbotPredictionsContext } from './types';
 
 const ajv = new Ajv();
-const validate_predictions = ajv.compile<PredictionsContext>(predictions_schema);
+const validate_predictions = ajv.compile<EzbotPredictionsContext>(
+  predictionsContextSchema
+);
 
 const predictions = [
   {
@@ -100,7 +98,12 @@ describe('ezbot js tracker', () => {
     const contexts = firstEvent.evt.cx;
     const decodedContexts = decodeContexts(contexts as string);
     expect(decodedContexts).toContainEqual({
-      data: { predictions: predictions.map(p => ({variable: p.key, value: p.value})) },
+      data: {
+        predictions: predictions.map((p) => ({
+          variable: p.key,
+          value: p.value,
+        })),
+      },
       schema: 'iglu:com.ezbot/predictions_context/jsonschema/1-0-1',
     });
   });
@@ -111,7 +114,7 @@ describe('ezbot js tracker', () => {
     const contexts = firstEvent.evt.cx;
     const decodedContexts: Context[] = decodeContexts(contexts as string);
 
-    expect(validate_predictions(decodedContexts[2].data)).toBeTruthy()
+    expect(validate_predictions(decodedContexts[2].data)).toBeTruthy();
   });
   it('exposes a global trackPageView function', async () => {
     expect(tracker.trackPageView).toBeDefined();
